requirements.txt
# Updated requirements.txt - Windows-compatible with NumPy 2.x
torch>=2.6.0
torchvision>=0.21.0
pytorch-lightning>=2.5.0
numpy>=2.1.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
mediapipe>=0.10.0
scikit-learn>=1.3.0
scipy>=1.11.0
pyyaml>=6.0
tensorboard>=2.15.0
pandas>=2.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
Pillow>=10.0.0